rafactored file structures and write test bootstrap
